+++ CarService.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace SOLID_violation
+{
+    public class CarService
+    {
+        public void Wash(Car c) { }
+
+        public void SaveCarToDB(Car c)
+        {
+            // saving to DB
+        }
+    }
+}
