+++ ICar.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace SOLID_violation
+{
+    public interface ICar
+    {
+
+    }
+}
